+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/pages/administration-new/grid-col-def.ts
@@ -1,0 +1,92 @@
+import { GridColDef } from '@mui/x-data-grid';
+
+const dataColumns: GridColDef[] = [
+    {
+        field: 'id',
+        headerName: 'Id',
+        type: 'number',
+        filterable: true,
+        sortable: true,
+        valueFormatter: (params) => params.value.toString(),
+    },
+    {
+        field: 'problem',
+        headerName: 'Problem',
+        type: 'string',
+        filterable: true,
+        sortable: true,
+    },
+    {
+        field: 'problemId',
+        headerName: 'Problem Id',
+        type: 'number',
+        filterable: true,
+        sortable: true,
+    },
+    {
+        field: 'processed',
+        headerName: 'Processed',
+        align: 'left',
+        type: 'boolean',
+        filterable: true,
+        sortable: false,
+    },
+    {
+        field: 'points',
+        headerName: 'Points',
+        align: 'left',
+        type: 'number',
+        filterable: true,
+        sortable: true,
+    },
+    {
+        field: 'isCompiledSuccessfully',
+        headerName: 'Is Compiled Successfully',
+        align: 'left',
+        type: 'boolean',
+        filterable: true,
+        sortable: false,
+    },
+    {
+        field: 'processingComment',
+        headerName: 'Processing Comment',
+        align: 'left',
+        type: 'string',
+        filterable: true,
+        sortable: false,
+    },
+    {
+        field: 'submissionType',
+        headerName: 'Submission Type',
+        align: 'left',
+        type: 'string',
+        filterable: true,
+        sortable: true,
+    },
+    {
+        field: 'participantId',
+        headerName: 'Participant',
+        align: 'left',
+        type: 'number',
+        filterable: true,
+        sortable: false,
+    },
+    {
+        field: 'participant',
+        headerName: 'Participant',
+        align: 'left',
+        type: 'string',
+        filterable: true,
+        sortable: false,
+    },
+    {
+        field: 'isDeleted',
+        headerName: 'Is Deleted',
+        align: 'left',
+        type: 'date',
+        filterable: true,
+        sortable: false,
+    },
+];
+
+export default dataColumns;
